requirement.txt
selenium==3.11.0
html5lib==0.9999999
PyExecJS==1.5.1
urllib3==1.22
bs4==0.0.1

